+++ pallas-hardano/Cargo.toml
@@ -1,9 +1,14 @@
 [package]
 name = "pallas-hardano"
-version = "0.1.0"
+description = "Pallas interoperability with the Haskel Cardano node implementation"
+version = "0.19.1"
 edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+repository = "https://github.com/txpipe/pallas"
+homepage = "https://github.com/txpipe/pallas"
+documentation = "https://docs.rs/pallas-hardano"
+license = "Apache-2.0"
+readme = "README.md"
+authors = ["Santiago Carmuega <santiago@carmuega.me>"]
 
 [dependencies]
 binary-layout = "3.2.0"
