+++ pallas-hardano/src/storage/immutable/chunk.rs
@@ -4,7 +4,10 @@ use std::{
     path::Path,
 };
 
+use immutable::secondary;
 use tracing::trace;
+
+use crate::storage::immutable;
 
 pub type SecondaryIndex = super::secondary::Reader;
 
@@ -91,11 +94,11 @@ impl Iterator for Reader {
 pub fn read_blocks(dir: &Path, name: &str) -> Result<Reader, std::io::Error> {
     let primary = dir.join(name).with_extension("primary");
     let primary = std::fs::File::open(&primary)?;
-    let primary = crate::primary::Reader::open(primary)?;
+    let primary = immutable::primary::Reader::open(primary)?;
 
     let secondary = dir.join(name).with_extension("secondary");
     let secondary = std::fs::File::open(&secondary)?;
-    let secondary = crate::secondary::Reader::open(primary, secondary)?;
+    let secondary = secondary::Reader::open(primary, secondary)?;
 
     let chunk = dir.join(name).with_extension("chunk");
     let chunk = std::fs::File::open(&chunk)?;
