+++ pallas-hardano/src/storage/mod.rs
@@ -1,0 +1,3 @@
+//! Storage compatible with the Haskell Cardano node implementation
+
+pub mod immutable;
