+++ pallas/Cargo.toml
@@ -21,6 +21,7 @@ pallas-codec = { version = "=0.19.1", path = "../pallas-codec/" }
 pallas-utxorpc = { version = "=0.19.1", path = "../pallas-utxorpc/" }
 pallas-configs = { version = "=0.19.1", path = "../pallas-configs/" }
 pallas-rolldb = { version = "=0.19.1", path = "../pallas-rolldb/" }
+pallas-hardano = { version = "=0.19.1", path = "../pallas-hardano/", optional = true }
 
 [features]
-unstable = []
+unstable = ["pallas-hardano"]
